+++ monetrack/__init__.py
@@ -1,1 +1,1 @@
-# MonetRack Package
+# MoneTrack Package